feat: update with new api
I didn't find a way to filter favorites talks with this api, so I generated all talks and we'll need to only copy the one we need.

Reference : https://devoxxfr2024.cfp.dev/swagger-ui/index.html#/
