+++ README.md
@@ -2,8 +2,6 @@
 
 Generate your pre-filled notes in adoc and just add your notes during presentations !
 
-This script only generate the talks you marked as favorite in your planning : <https://cfp.devoxx.fr/2023/byday/wed>
-
 ## Requirements
 
 ```bash
@@ -15,7 +13,9 @@ pip3 install -r requirements.txt --user
 Get the cookie value of `PLAY_SESSION` in the `https://cfp.devoxx.fr` domain.
 
 ```bash
-./favorite.py PLAY_SESSION_VALUE
+./generate.py
 ```
 
-This will generate a folder `wednesday`, `thursday` and `friday` with your favorite talks, based on the template `template.adoc`.
+This will generate a folder `talks` with all talks, based on the template `template.adoc`.
+
+Pick your favorites and add your notes !
