+++ requirements.txt
@@ -1,1 +1,2 @@
 requests==2.28.1
+html2markdown==0.1.7
